+++ packages/ice-plugin-component/lib/compile/component/buildSrc.js
@@ -5,12 +5,13 @@
  *    - 生成 style.js 和 index.scss
  */
 
-const { createReadStream, createWriteStream, writeFileSync } = require('fs');
+const { createReadStream, createWriteStream, writeFileSync, ensureDirSync } = require('fs-extra');
 const babel = require('@babel/core');
 const glob = require('glob');
 const mkdirp = require('mkdirp');
 const path = require('path');
 const rimraf = require('rimraf');
+const ts = require('typescript');
 
 module.exports = function componentBuild({ babelConfig, rootDir, log }) {
   const srcDir = path.join(rootDir, 'src');
@@ -54,6 +55,7 @@ module.exports = function componentBuild({ babelConfig, rootDir, log }) {
       filename: file,
     }));
     writeFileSync(path.format(destData), code, 'utf-8');
+    dtsCompile({ filePath: file, sourceFile: source, destPath: libDir });
     log.info(`Compile ${file}`);
   }
 
@@ -68,5 +70,35 @@ module.exports = function componentBuild({ babelConfig, rootDir, log }) {
       .on('close', () => {
         log.info(`Copy ${file}`);
       });
+  }
+  // https://github.com/microsoft/TypeScript/wiki/Using-the-Compiler-API#getting-the-dts-from-a-javascript-file
+  function dtsCompile({ filePath, sourceFile, destPath }) {
+    const REG_TS = /\.(tsx?)$/;
+    const isTS = REG_TS.test(filePath);
+    if (!isTS) return;
+    const compilerOptions = {
+      allowJs: true,
+      declaration: true,
+      emitDeclarationOnly: true,
+    };
+    const dtsPath = filePath.replace(REG_TS, '.d.ts');
+    const targetPath = path.join(destPath, dtsPath);
+    // Create a Program with an in-memory emit
+    let createdFiles = {};
+    const host = ts.createCompilerHost(compilerOptions);
+    host.writeFile = (fileName, contents) => createdFiles[fileName] = contents;
+    // Prepare and emit the d.ts files
+    const program = ts.createProgram([sourceFile], compilerOptions, host);
+    program.emit();
+    const fileNamesDTS = sourceFile.replace(REG_TS, '.d.ts');
+    const content = createdFiles[fileNamesDTS];
+    // write file
+    if (content) {
+      ensureDirSync(path.dirname(targetPath));
+      writeFileSync(targetPath, content, 'utf-8');
+      log.info(`Generate ${path.basename(targetPath)}`);
+    }
+    // release
+    createdFiles = null;
   }
 };