+++ packages/ice-plugin-component/lib/configs/build.js
@@ -15,9 +15,10 @@ module.exports = (config, { demos, markdownParser, rootDir }) => {
     }),
   };
   config.output.publicPath('./');
-  ['scss', 'scss-module', 'css', 'css-module', 'less', 'less-module'].forEach((rule) => {
+  ['scss', 'scss-module', 'css', 'css-module', 'less', 'less-module', 'styl', 'styl-module'].forEach((rule) => {
     if (config.module.rules.get(rule)) {
-      config.module.rule(rule).use('MiniCssExtractPlugin.loader').tap(() => ({ publicPath: '../' }));
+      // css output path is '', set publicPath './'
+      config.module.rule(rule).use('MiniCssExtractPlugin.loader').tap(() => ({ publicPath: './' }));
     }
   });
   // add hbs loader