+++ packages/ice-scripts/bin/ice-scripts-build.js
@@ -6,6 +6,9 @@ const log = require('../lib/utils/log');
 
 program
   .option('--config <config>', 'use custom config')
+  .option('--analyzer', '开启构建分析')
+  .option('--analyzer-port', '设置分析端口号')
+  .option('--skip-compile', 'skip webpack compile, excute hooks for component compile')
   .parse(process.argv);
 
 (async () => {