+++ packages/ice-scripts/bin/ice-scripts-test.js
@@ -1,9 +1,10 @@
 #!/usr/bin/env node
 const program = require('commander');
+const parse = require('yargs-parser');
+
 const getCliOptions = require('../lib/utils/getCliOptions');
 const Context = require('../lib/core/Context');
 const log = require('../lib/utils/log');
-const parse = require('yargs-parser');
 
 program
   .allowUnknownOption() // allow jest config
@@ -23,6 +24,9 @@ program
       jestArgv[`${jestKey[0].toLowerCase()}${jestKey.slice(1)}`] = rawArgv[key];
     }
   });
+  if (program.args && program.args.length > 0) {
+    jestArgv.regexForTestFiles = program.args;
+  }
   try {
     await new Context({
       command: 'test',