+++ packages/ice-scripts/lib/commands/dev.js
@@ -14,24 +14,18 @@ const WebpackDevServer = require('webpack-dev-server');
 const openBrowser = require('react-dev-utils/openBrowser');
 const iceworksClient = require('../utils/iceworksClient');
 const prepareUrLs = require('../utils/prepareURLs');
-const goldlog = require('../utils/goldlog');
 const pkgData = require('../../package.json');
 const log = require('../utils/log');
 const checkDepsInstalled = require('../utils/checkDepsInstalled');
 
-module.exports = async function (context, subprocess) {
+module.exports = async function(context, subprocess) {
   const { applyHook, commandArgs, rootDir, webpackConfig, pkg } = context;
-
-  goldlog('version', {
-    version: pkgData.version,
-  });
-  goldlog('dev', commandArgs);
   log.verbose('dev cliOptions', commandArgs);
 
   await applyHook('beforeDev');
 
   // 与 iceworks 客户端通信
-  const send = function (data) {
+  const send = function(data) {
     iceworksClient.send(data);
     if (subprocess && typeof subprocess.send === 'function') {
       subprocess.send(data);
@@ -43,14 +37,14 @@ module.exports = async function (context, subprocess) {
     return Promise.reject(new Error('项目依赖未安装，请先安装依赖。'));
   }
 
-  if (!pkg.componentConfig && !pkg.blockConfig) {
+  if (!pkg.componentConfig && !pkg.blockConfig && !process.env.DISABLE_COLLECT) {
     // only collect project
     try {
       collectDetail({
         rootDir,
         kit: 'ice-scripts',
         kitVersion: pkgData.version,
-        cmd_type: 'dev',
+        cmdType: process.stderr.isTTY ? 'dev' : 'nontty-dev',
       });
     } catch (err) {
       log.warn('collectDetail error', err);
@@ -69,13 +63,26 @@ module.exports = async function (context, subprocess) {
   }
 
   let isFirstCompile = true;
+
+  if (commandArgs.disabledMock) {
+    log.warn('关闭了 mock 功能');
+  } else {
+    // devMock 在 devServer 之前启动
+    const originalDevServeBefore = webpackConfig.devServer.before;
+    webpackConfig.devServer.before = function(app, server) {
+      // dev mock
+      webpackDevMock(app);
+
+      if (typeof originalDevServeBefore === 'function') {
+        originalDevServeBefore(app, server);
+      }
+    };
+  }
+
   const compiler = webpack(webpackConfig);
   const devServer = new WebpackDevServer(compiler, webpackConfig.devServer);
 
-  // dev mock
-  webpackDevMock(devServer.app);
-
-  compiler.hooks.done.tap('done', (stats) => {
+  compiler.hooks.done.tap('done', stats => {
     if (isInteractive) {
       clearConsole();
     }
@@ -97,7 +104,9 @@ module.exports = async function (context, subprocess) {
           `    - Network: ${chalk.yellow(urls.lanUrlForTerminal)}`,
         ].join('\n')
       );
-      openBrowser(urls.localUrlForBrowser);
+      if (!commandArgs.disabledOpen) {
+        openBrowser(urls.localUrlForBrowser);
+      }
     }
 
     applyHook('afterDev', stats);
@@ -120,9 +129,7 @@ module.exports = async function (context, subprocess) {
       if (stats.stats) {
         log.info('Compiled successfully');
       } else {
-        log.info(
-          `Compiled successfully in ${(json.time / 1000).toFixed(1)}s!`
-        );
+        log.info(`Compiled successfully in ${(json.time / 1000).toFixed(1)}s!`);
       }
 
       // 服务启动完成切没有任务错误与警告
@@ -143,21 +150,17 @@ module.exports = async function (context, subprocess) {
       } else if (messages.warnings.length) {
         log.warn('Compiled with warnings.');
         console.log();
-        messages.warnings.forEach((message) => {
+        messages.warnings.forEach(message => {
           console.log(message);
           console.log();
         });
         // Teach some ESLint tricks.
         console.log('You may use special comments to disable some warnings.');
         console.log(
-          `Use ${chalk.yellow(
-            '// eslint-disable-next-line'
-          )} to ignore the next line.`
+          `Use ${chalk.yellow('// eslint-disable-next-line')} to ignore the next line.`
         );
         console.log(
-          `Use ${chalk.yellow(
-            '/* eslint-disable */'
-          )} to ignore all warnings in a file.`
+          `Use ${chalk.yellow('/* eslint-disable */')} to ignore all warnings in a file.`
         );
         console.log();
       }
@@ -193,7 +196,7 @@ module.exports = async function (context, subprocess) {
     next();
   });
 
-  devServer.listen(PORT, HOST, (err) => {
+  devServer.listen(PORT, HOST, err => {
     if (err) {
       send({
         action: 'update_project',