+++ packages/ice-scripts/lib/commands/test.js
@@ -4,10 +4,10 @@ const jest = require('jest');
 
 module.exports = (context) => {
   const { commandArgs: { jestArgv }, rootDir, webpackConfig } = context;
-
+  const { config, regexForTestFiles, ...restArgv } = jestArgv;
   // get user jest config
-  const jestConfigPath = jestArgv.config
-    ? path.join(rootDir, jestArgv.config)
+  const jestConfigPath = config
+    ? path.join(rootDir, config)
     : path.join(rootDir, 'jest.config.js');
   let userJestConfig = {};
   if (fs.existsSync(jestConfigPath)) {
@@ -29,7 +29,7 @@ module.exports = (context) => {
   // generate default jest config
   const jestConfig = {
     rootDir,
-    setupFiles: [require.resolve('@babel/polyfill')],
+    setupFiles: [require.resolve('../config/jest/shim.js')],
     testMatch: ['**/?*.(spec|test).(j|t)s?(x)'],
     transform: {
       '^.+\\.(js|jsx|ts|tsx)$': require.resolve('../config/jest/babelTransform.js'),
@@ -47,12 +47,13 @@ module.exports = (context) => {
     moduleFileExtensions: ['js', 'jsx', 'ts', 'tsx', 'json'],
     testPathIgnorePatterns: ['/node_modules/'],
     ...userJestConfig,
+    ...(regexForTestFiles ? { testMatch: regexForTestFiles } : {}),
   };
 
   return new Promise((resolve, reject) => {
     jest.runCLI(
       {
-        ...jestArgv,
+        ...restArgv,
         config: JSON.stringify(jestConfig),
       },
       [rootDir],