+++ packages/ice-scripts/lib/config/default.config.js
@@ -44,4 +44,5 @@ module.exports = {
   publicPath: '/',
   targets: 'last 2 versions, Firefox ESR, > 1%, ie >= 9, iOS >= 8, Android >= 4',
   vendor: true,
+  postcssrc: false,
 };