+++ packages/ice-scripts/lib/config/jest/babelTransform.js
@@ -1,10 +1,11 @@
 const babelJest = require('babel-jest');
 const getBabelConfig = require('../getBabelConfig');
+const formatWinPath = require('../../utils/formatWinPath');
 
 const babelConfig = getBabelConfig();
 // options for @babel/preset-env when test
 babelConfig.presets = babelConfig.presets.map((preset) => {
-  if (Array.isArray(preset) && preset[0].indexOf('@babel/preset-env') > -1) {
+  if (Array.isArray(preset) && formatWinPath(preset[0]).indexOf('@babel/preset-env') > -1) {
     return [preset[0], {
       targets: {
         node: 'current',