+++ packages/ice-plugin-fusion/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.1.12
+
+- [feat] support option nextLibDir for change libraryDirectory
+
+## 0.1.11
+
+- [fix] problem with importing scss file paths on windows
+
+## 0.1.10
+
+- [feat] 多主题支持 @alifd/next 组件色值计算 sass 变量替换
+
+## 0.1.9
+
+- [fix] 多主题切换函数 `__changeTheme__` 前置注入
+
+## 0.1.8
+
+- [fix] 修复 1.x 组件主题 icon 编译问题
+
 ## 0.1.7
 
 - [feat] themePackage 支持多主题切换