+++ packages/ice-plugin-smart-debug/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.2
+
+ - [fix] post plugin api for wrap code
+  
 ## 0.1.1
 
  - [fix] 通过 smartDebug 开启调试服务